let only one addResourceForEntity method, moved UnitOfWorkEventos to core, surefireArgLine is back
