+++ crabzilla-pg-client-jooq/src/test/java/io/github/crabzilla/pgc/jooq/PgcJooqUowProjectorIT.kt
@@ -2,7 +2,7 @@ package io.github.crabzilla.pgc.jooq
 
 import io.github.crabzilla.core.DomainEvent
 import io.github.crabzilla.core.UnitOfWork
-import io.github.crabzilla.internal.UnitOfWorkEvents
+import io.github.crabzilla.core.UnitOfWorkEvents
 import io.github.crabzilla.pgc.example1.CustomerActivated
 import io.github.crabzilla.pgc.example1.CustomerCreated
 import io.github.crabzilla.pgc.example1.CustomerDeactivated
@@ -33,12 +33,10 @@ import org.jooq.impl.DefaultConfiguration
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.extension.ExtendWith
 import org.slf4j.LoggerFactory
 
 @ExtendWith(VertxExtension::class)
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class PgcJooqUowProjectorIT {
 
   private lateinit var vertx: Vertx
@@ -108,6 +106,7 @@ class PgcJooqUowProjectorIT {
             tc.failNow(deleted2.cause())
             return@execute
           }
+          println("** db cleaned")
           tc.completeNow()
         }
       }
